Renames layouts folder, layout component, icons folder
