+++ src/components/icons/Calendar.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgCalendar = (props: SVGProps<SVGSVGElement>) => (
-  <svg xmlns="http://www.w3.org/2000/svg" width={19} height={22} fill="none" {...props}>
+  <svg xmlns="http://www.w3.org/2000/svg" width={19} height={24} fill="none" {...props}>
     <path
       fill="currentColor"
       d="M13.076 2.396h-.204v-.204a.612.612 0 0 0-1.223 0v.204H6.351v-.204a.611.611 0 1 0-1.223 0v.204h-.204A3.877 3.877 0 0 0 1.052 6.27v7.336a3.877 3.877 0 0 0 3.872 3.873h8.152a3.877 3.877 0 0 0 3.872-3.873V6.27a3.877 3.877 0 0 0-3.872-3.873ZM4.924 3.62h.204v1.02a.611.611 0 1 0 1.223 0v-1.02h5.298v1.02a.612.612 0 0 0 1.223 0v-1.02h.204a2.653 2.653 0 0 1 2.65 2.65v.203H2.275V6.27a2.653 2.653 0 0 1 2.649-2.65Zm8.152 12.636H4.924a2.653 2.653 0 0 1-2.65-2.65v-5.91h13.452v5.91a2.653 2.653 0 0 1-2.65 2.65Zm0-5.91a.815.815 0 1 1-1.63 0 .815.815 0 0 1 1.63 0Zm-3.26 0a.815.815 0 1 1-1.631 0 .815.815 0 0 1 1.63 0Zm-3.262 0a.815.815 0 1 1-1.63 0 .815.815 0 0 1 1.63 0Zm6.522 3.26a.815.815 0 1 1-1.63 0 .815.815 0 0 1 1.63 0Zm-3.26 0a.815.815 0 1 1-1.63 0 .815.815 0 0 1 1.63 0Zm-3.262 0a.815.815 0 1 1-1.63 0 .815.815 0 0 1 1.63 0Z"
