+++ src/components/icons/Image.tsx
