+++ src/layouts/Sidebebar/Sidebar.tsx
@@ -2,7 +2,7 @@ import { Flex, Text } from '@chakra-ui/react';
 import Image from 'next/image';
 import { useRouter } from 'next/router';
 
-import { CalendarIcon, CardIcon, ComputerIcon, ImageIcon } from '@/components/Icons';
+import { CalendarIcon, CardIcon, ComputerIcon, ImageIcon } from '@/components/icons';
 
 import { MenuItem } from './MenuItem';
